+++ src/pages/components/content item/LinksSection.tsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom"
 import { LinksType } from "../../../data/types"
-import { AUnderline, LinkUnderline } from "../links"
+import { AUnderline, LinkUnderline } from "../underline links"
 
 export const LinksSection = (links?: LinksType) => {
     if (!links) return <></>
